+++ src/Translators/GoogleTranslateLegitimate/GoogleTranslateLegitimateEndpoint.cs
@@ -19,7 +19,7 @@ namespace GoogleTranslateLegitimate
 {
    internal class GoogleTranslateLegitimateEndpoint : HttpEndpoint
    {
-      private static readonly HashSet<string> SupportedLanguages = new HashSet<string> { "af","sq","am","ar","hy","az","eu","be","bn","bs","bg","ca","ceb","zh-CN","zh-TW","co","hr","cs","da","nl","en","eo","et","fi","fr","fy","gl","ka","de","el","gu","ht","ha","haw","he","hi","hmn","hu","is","ig","id","ga","it","ja","jw","kn","kk","km","ko","ku","ky","lo","la","lv","lt","lb","mk","mg","ms","ml","mt","mi","mr","mn","my","ne","no","ny","ps","fa","pl","pt","pa","ro","ru","sm","gd","sr","st","sn","sd","si","sk","sl","so","es","su","sw","sv","tl","tg","ta","te","th","tr","uk","ur","uz","vi","cy","xh","yi","yo","zu" };
+      private static readonly HashSet<string> SupportedLanguages = new HashSet<string> { "af","sq","am","ar","hy","az","eu","be","bn","bs","bg","ca","ceb","zh","zh-Hans","zh-Hant","zh-CN","zh-TW","co","hr","cs","da","nl","en","eo","et","fi","fr","fy","gl","ka","de","el","gu","ht","ha","haw","he","hi","hmn","hu","is","ig","id","ga","it","ja","jw","kn","kk","km","ko","ku","ky","lo","la","lv","lt","lb","mk","mg","ms","ml","mt","mi","mr","mn","my","ne","no","ny","ps","fa","pl","pt","pa","ro","ru","sm","gd","sr","st","sn","sd","si","sk","sl","so","es","su","sw","sv","tl","tg","ta","te","th","tr","uk","ur","uz","vi","cy","xh","yi","yo","zu" };
       private static readonly string HttpsServiceUrl = "https://translation.googleapis.com/language/translate/v2";
 
       private string _key;
@@ -30,6 +30,20 @@ namespace GoogleTranslateLegitimate
 
       public override int MaxTranslationsPerRequest => 10;
 
+      private string FixLanguage( string lang )
+      {
+         switch( lang )
+         {
+            case "zh-Hans":
+            case "zh":
+               return "zh-CN";
+            case "zh-Hant":
+               return "zh-TW";
+            default:
+               return lang;
+         }
+      }
+
       public override void Initialize( IInitializationContext context )
       {
          _key = context.GetOrCreateSetting( "GoogleLegitimate", "GoogleAPIKey", "" );
@@ -38,16 +52,16 @@ namespace GoogleTranslateLegitimate
          // Configure service points / service point manager
          context.DisableCertificateChecksFor( "translation.googleapis.com" );
 
-         if( !SupportedLanguages.Contains( context.SourceLanguage ) ) throw new EndpointInitializationException( $"The source language '{context.SourceLanguage}' is not supported." );
-         if( !SupportedLanguages.Contains( context.DestinationLanguage ) ) throw new EndpointInitializationException( $"The destination language '{context.DestinationLanguage}' is not supported." );
+         if( !SupportedLanguages.Contains( FixLanguage( context.SourceLanguage ) ) ) throw new EndpointInitializationException( $"The source language '{context.SourceLanguage}' is not supported." );
+         if( !SupportedLanguages.Contains( FixLanguage( context.DestinationLanguage ) ) ) throw new EndpointInitializationException( $"The destination language '{context.DestinationLanguage}' is not supported." );
       }
 
       public override void OnCreateRequest( IHttpRequestCreationContext context )
       {
          var urlBuilder = new StringBuilder( HttpsServiceUrl );
          urlBuilder.Append( "?key=" ).Append( Uri.EscapeDataString( _key ) );
-         urlBuilder.Append( "&source=" ).Append( context.SourceLanguage );
-         urlBuilder.Append( "&target=" ).Append( context.DestinationLanguage );
+         urlBuilder.Append( "&source=" ).Append( FixLanguage( context.SourceLanguage ) );
+         urlBuilder.Append( "&target=" ).Append( FixLanguage( context.DestinationLanguage ) );
          for( int i = 0 ; i < context.UntranslatedTexts.Length ; i++ )
          {
             var untranslatedText = context.UntranslatedTexts[ i ];