+++ src/XUnity.AutoTranslator.Plugin.Core/Configuration/Settings.cs
@@ -126,6 +126,7 @@ namespace XUnity.AutoTranslator.Plugin.Core.Configuration
       public static bool EnableLegacyTextureLoading;
       public static HashSet<string> DuplicateTextureNames;
       public static TextureHashGenerationStrategy TextureHashGenerationStrategy;
+      public static bool CacheTexturesInMemory;
 
       public static string PreferredStoragePath;
       public static bool EnableDumping;
@@ -220,7 +221,9 @@ namespace XUnity.AutoTranslator.Plugin.Core.Configuration
                ?.Split( new[] { ';' }, StringSplitOptions.RemoveEmptyEntries ).ToHashSet() ?? new HashSet<string>();
             EnableLegacyTextureLoading = PluginEnvironment.Current.Preferences.GetOrDefault( "Texture", "EnableLegacyTextureLoading", false );
             TextureHashGenerationStrategy = PluginEnvironment.Current.Preferences.GetOrDefault( "Texture", "TextureHashGenerationStrategy", TextureHashGenerationStrategy.FromImageName );
+            CacheTexturesInMemory = PluginEnvironment.Current.Preferences.GetOrDefault( "Texture", "CacheTexturesInMemory", true );
 
+            
             PreferredStoragePath = PluginEnvironment.Current.Preferences.GetOrDefault( "ResourceRedirector", "PreferredStoragePath", @"Translation\{Lang}\RedirectedResources" );
             EnableTextAssetRedirector = PluginEnvironment.Current.Preferences.GetOrDefault( "ResourceRedirector", "EnableTextAssetRedirector", false );
             LogAllLoadedResources = PluginEnvironment.Current.Preferences.GetOrDefault( "ResourceRedirector", "LogAllLoadedResources", false );